LayoutTests/http/tests/media/video-load-twice-expected.txt

RUN(video.src = file)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
RUN(video.src = file)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
END OF TEST
